GOV-1105: Move custom matchers to shared-test module
- Move the `AuditMessageMatcher` to the `shared-test` module
- Move the `APIGatewayProxyResponseEventMatcher` to the `shared-test` module
