+++ client-registry-api/build.gradle
@@ -20,7 +20,8 @@ dependencies {
     runtimeOnly configurations.logging_runtime
 
     testImplementation configurations.tests,
-            configurations.lambda_tests
+            configurations.lambda_tests,
+            project(":shared-test")
 
     testRuntimeOnly configurations.test_runtime
 }
