+++ shared-test/build.gradle
@@ -13,7 +13,9 @@ dependencies {
     implementation configurations.tests,
             configurations.lambda,
             configurations.glassfish,
-            "org.eclipse.jetty:jetty-server:11.0.7"
+            "org.eclipse.jetty:jetty-server:11.0.7",
+            "com.google.protobuf:protobuf-java:3.18.1",
+            project(":shared")
 }
 
 test {
