+++ shared/build.gradle
@@ -26,7 +26,8 @@ dependencies {
             "com.google.protobuf:protobuf-java:3.18.1"
 
     testImplementation configurations.tests,
-            configurations.lambda_tests
+            configurations.lambda_tests,
+            project(":shared-test")
     testRuntimeOnly configurations.test_runtime
 }
